# perf(types): 3x speedup MakePartSet (#3117)